Add unit test to counters and card generation function
