+++ src/index.html
@@ -14,16 +14,14 @@
     <nav>
         <img src="./assets/logo-no-background.svg" alt="logo" id="h-logo">
         <ul>
-            <li>Most Trendy (6)</li>
-            <li>Best Rated</li>
-            <li>Most Viewed</li>
+            <li>Most Trendy <p id="mCount"></p></li>
         </ul>  
     </nav>
 </header>
 <main>
     <div class="w-75 mx-auto">
         <div class="section-title">
-            <h2>Most Trendy Movies:</h2>
+            <h2>The latest trends:</h2>
         </div>
         <div class="row .movies-grid" id="moviesContainer">
         </div>
