+++ src/main.scss
@@ -69,7 +69,7 @@ header {
         list-style: none;
         text-transform: uppercase;
         font-weight: 600;
-        font-size: 2rem;
+        font-size: 1.5rem;
         padding: 1rem 1.5rem;
       }
     }
